+++ src/main/kotlin/de/samthedev/foliapapiexpansion/PlatformDetector.kt
@@ -2,7 +2,7 @@
 // Copyright (C) 2026 SamTheDevDE
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-package de.samthedev.FoliaPAPIExpansion
+package de.samthedev.foliapapiexpansion
 
 import io.papermc.paper.ServerBuildInfo
 import net.kyori.adventure.key.Key