+++ src/main/kotlin/de/samthedev/foliapapiexpansion/health/HealthStatus.kt
@@ -2,7 +2,7 @@
 // Copyright (C) 2026 SamTheDevDE
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-package de.samthedev.FoliaPAPIExpansion.health
+package de.samthedev.foliapapiexpansion.health
 
 internal enum class HealthStatus {
     HEALTHY,